backend/gemini.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

# Load .env file
load_dotenv()

# Set API key from .env
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

# Load the Gemini model
model = genai.GenerativeModel(model_name="models/gemini-2.5-pro")

def generate_reply(prompt):
    try:
        response = model.generate_content(prompt)
        return response.text
    except Exception as e:
        print("❌ Gemini error:", e)
        return "Sorry, I couldn’t generate a reply at the moment."
